requirements.txt
flask-sqlalchemy
flask-migrate
pymysqlS